scheduler: Apply robustness and safety improvements from 2025 audit
This patch applies several critical improvements to the Haiku scheduler
identified during the 2025 audit:

*   Add node index bounds checks in `choose_core` (low_latency.cpp),
    `choose_idle_core` (power_saving.cpp), and `dump_idle_cores`
    (scheduler_cpu.cpp) to prevent out-of-bounds access to
    `gSchedulerNodes`.
*   Ensure bit-shift and bitwise operations in `scheduler_common.h` and
    `scheduler_cpu.h` are 64-bit safe by using explicit `uint64` casts
    before shifting, preventing signed overflow issues.
*   Add a NULL check for `CPUEntry` in `CheckMaskedPackagesMinimumLoad`
    (scheduler_topology.h) to ensure stability during CPU hot-unplug.
*   Modify `RemoveThread::Name()` (scheduler_tracing.cpp) to return an
    empty string instead of `NULL` for consistent trace handling.
*   Improve indentation for consistency in modified blocks.

+++ src/system/kernel/scheduler/scheduler_common.h
@@ -476,19 +476,19 @@ inline void AssertThreadQueued(Thread* thread) {
 inline void SetCPUIDle(uint64 volatile& mask, int cpu) {
 	if ((unsigned)cpu >= 64)
 		return;
-	OrAtomic64(mask, (int64)1 << cpu);
+	OrAtomic64(mask, (int64)((uint64)1 << cpu));
 }
 
 inline void ClearCPUIDle(uint64 volatile& mask, int cpu) {
 	if ((unsigned)cpu >= 64)
 		return;
-	AndAtomic64(mask, ~((int64)1 << cpu));
+	AndAtomic64(mask, (int64)~((uint64)1 << cpu));
 }
 
 inline bool IsCPUIDle(const uint64 volatile& mask, int cpu) {
 	if ((unsigned)cpu >= 64)
 		return false;
-	return (LoadAcquire64(mask) & ((int64)1 << cpu)) != 0;
+	return (LoadAcquire64(mask) & ((uint64)1 << cpu)) != 0;
 }
 
 struct SchedulerSnapshot {
